Updated Nilmod to 0.8.0.5
- Removed husk hide list
- Refactored MultiPlayerCampaign to reduce code repetition with methods

+++ Barotrauma/BarotraumaShared/Source/Characters/CharacterNetworking.cs
@@ -405,7 +405,7 @@ namespace Barotrauma
                             }
                             break;
                         case 3:
-                            LimbType grabLimb = (LimbType)msg.ReadUInt16();
+                            LimbType grabLimb = (LimbType)msg.ReadByte();
                             if (c.Character != this)
                             {
 #if DEBUG
@@ -496,7 +496,7 @@ namespace Barotrauma
                     if (AnimController is HumanoidAnimController)
                     {
                         tempBuffer.Write(((HumanoidAnimController)AnimController).Crouching);
-                        tempBuffer.Write((UInt16)AnimController.GrabLimb);
+                        tempBuffer.Write((byte)AnimController.GrabLimb);
                     }
 
                     bool hasAttackLimb = AnimController.Limbs.Any(l => l != null && l.attack != null);
